+++ eslint.config.mjs
@@ -1,21 +1,22 @@
-import eslint from "@eslint/js"
-import tseslint from "typescript-eslint"
+import js from "@eslint/js"
+import pluginChaiFriendly from "eslint-plugin-chai-friendly"
+import { jsdoc } from "eslint-plugin-jsdoc"
+import { defineConfig, globalIgnores } from "eslint/config"
 import globals from "globals"
+import ts from "typescript-eslint"
 
-export default tseslint.config([
-    {
-        ignores: [
-            "build/**",
-            "docs/**",
-            "node_modules/**",
-            "sample/playground/**",
-            "temp/**",
-        ],
-    },
+export default defineConfig([
+    globalIgnores([
+        "build/**",
+        "docs/**",
+        "node_modules/**",
+        "sample/playground/**",
+        "temp/**",
+    ]),
     {
         files: ["**/*.ts"],
         languageOptions: {
-            parser: tseslint.parser,
+            parser: ts.parser,
             parserOptions: {
                 project: "tsconfig.json",
             },
@@ -24,10 +25,11 @@ export default tseslint.config([
                 ...globals.node,
             },
         },
-        extends: [
-            eslint.configs.recommended,
-            ...tseslint.configs.recommendedTypeChecked,
-        ],
+        plugins: {
+            js,
+            ts,
+        },
+        extends: [js.configs.recommended, ...ts.configs.recommendedTypeChecked],
         rules: {
             // exceptions from typescript-eslint/recommended
             "@typescript-eslint/ban-ts-comment": "warn",
@@ -38,10 +40,12 @@ export default tseslint.config([
             "@typescript-eslint/no-unnecessary-type-constraint": "warn",
             "@typescript-eslint/no-unsafe-declaration-merging": "warn",
             "@typescript-eslint/no-unsafe-function-type": "warn",
-            "@typescript-eslint/no-unused-expressions": "warn",
             "@typescript-eslint/no-unused-vars": [
                 "warn",
-                { argsIgnorePattern: "^_" },
+                {
+                    argsIgnorePattern: "^_",
+                    destructuredArrayIgnorePattern: "^_",
+                },
             ],
             "@typescript-eslint/no-wrapper-object-types": "off",
             "prefer-const": ["error", { destructuring: "all" }],
@@ -79,5 +83,18 @@ export default tseslint.config([
             "no-prototype-builtins": "warn",
             "no-regex-spaces": "warn",
         },
+    },
+    jsdoc({
+        files: ["src/**/*.ts"],
+        config: "flat/recommended-typescript",
+        // Temporarily enable individual rules when they are fixed, until all current warnings are gone,
+        // and then remove manual config in favor of `config: "flat/recommended-typescript-error"`
+        rules: {
+            "jsdoc/valid-types": "error",
+        },
+    }),
+    {
+        files: ["test/**/*.ts"],
+        ...pluginChaiFriendly.configs.recommendedFlat,
     },
 ])