+++ src/driver/types/ColumnTypes.ts
@@ -75,6 +75,8 @@ export type WithLengthColumnType =
     | "binary" // mssql
     | "varbinary" // mssql, sap
     | "string" // cockroachdb, spanner
+    | "vector" // mariadb, mysql, mssql, postgres, sap
+    | "halfvec" // postgres, sap
     | "half_vector" // sap
     | "real_vector" // sap
 
@@ -201,6 +203,7 @@ export type SimpleColumnType =
     | "xml" // mssql, postgres
     | "json" // mysql, postgres, cockroachdb, spanner
     | "jsonb" // postgres, cockroachdb
+    | "jsonpath" // postgres
     | "varbinary" // mssql, sap
     | "hierarchyid" // mssql
     | "sql_variant" // mssql