+++ src/query-builder/SelectQueryBuilder.ts
@@ -1872,9 +1872,8 @@ export class SelectQueryBuilder<Entity extends ObjectLiteral>
             }
 
             this.expressionMap.queryEntity = true
-            const entitiesAndRaw = await this.executeEntitiesAndRawResults(
-                queryRunner,
-            )
+            const entitiesAndRaw =
+                await this.executeEntitiesAndRawResults(queryRunner)
             this.expressionMap.queryEntity = false
 
             let count: number | undefined = this.lazyCount(entitiesAndRaw)
@@ -1929,8 +1928,8 @@ export class SelectQueryBuilder<Entity extends ObjectLiteral>
         const maxResults = hasLimit
             ? this.expressionMap.limit
             : hasTake
-            ? this.expressionMap.take
-            : undefined
+              ? this.expressionMap.take
+              : undefined
 
         if (
             maxResults !== undefined &&
@@ -1959,8 +1958,8 @@ export class SelectQueryBuilder<Entity extends ObjectLiteral>
         const previousResults: number = hasOffset
             ? this.expressionMap.offset!
             : hasSkip
-            ? this.expressionMap.skip!
-            : 0
+              ? this.expressionMap.skip!
+              : 0
 
         return entitiesAndRaw.entities.length + previousResults
     }
@@ -3621,9 +3620,8 @@ export class SelectQueryBuilder<Entity extends ObjectLiteral>
         if (rawResults.length > 0) {
             // transform raw results into entities
             const rawRelationIdResults = await relationIdLoader.load(rawResults)
-            const rawRelationCountResults = await relationCountLoader.load(
-                rawResults,
-            )
+            const rawRelationCountResults =
+                await relationCountLoader.load(rawResults)
             const transformer = new RawSqlResultsToEntityTransformer(
                 this.expressionMap,
                 this.connection.driver,
@@ -4215,8 +4213,8 @@ export class SelectQueryBuilder<Entity extends ObjectLiteral>
                     nulls?.toLowerCase() === "first"
                         ? "NULLS FIRST"
                         : nulls?.toLowerCase() === "last"
-                        ? "NULLS LAST"
-                        : undefined
+                          ? "NULLS LAST"
+                          : undefined
 
                 const aliasPath = `${alias}.${propertyPath}`
                 // const selection = this.expressionMap.selects.find(
@@ -4494,7 +4492,8 @@ export class SelectQueryBuilder<Entity extends ObjectLiteral>
                                             .inverseRelation!.inverseJoinColumns.map(
                                                 (column) => {
                                                     return `${
-                                                        relation.inverseRelation!
+                                                        relation
+                                                            .inverseRelation!
                                                             .joinTableName
                                                     }.${
                                                         column.propertyName