+++ src/schema-builder/options/TableIndexTypes.ts
@@ -1,0 +1,7 @@
+export type TableIndexTypes =
+    | "btree" // Postgres
+    | "hash" // Postgres
+    | "gist" // Postgres
+    | "spgist" // Postgres
+    | "gin" // Postgres
+    | "brin" // Postgres